students/k3343/practical_works/Tsybaeva_Arina/docker/eco_taxi/requirements.txt

appdirs==1.4.4
asgiref==3.2.5
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
DateTime==4.3
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
django-addresses==0.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-db==0.0.7
django-extensions==2.2.9
django-filter==2.2.0
django-filters==0.2.1
django-invoice==0.3.2
django-js-asset==1.2.2
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
entrypoints==0.3
filelock==3.0.12
future==0.18.2
graphviz==0.13.2
idna==2.9
inflection==0.4.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
openapi-codec==1.3.2
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2020.6.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psycopg2==2.8.5
public==2019.4.13
pycparser==2.20
pydotplus==2.0.2
Pygments==2.5.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
reportlab==3.5.42
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
timedelta==2019.4.13
tornado==6.0.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.23
virtualenv-clone==0.5.4
wcwidth==0.1.8
webencodings==0.5.1
zope.interface==5.1.0
